added slightly better error checking, fixed write end alignment
Write end alignment is now correct, and the error checking attempts to
warn the user if the SPI program operation failed.

chip_erase command on the MT35XU02GCBA does not seem to follow the
standard pattern for SPI flashes (where the command is sent, and no
bytes follow. Therefore, this command will fail on this flash since the
SR WEN bit will still be set, and we will fallthrough to chip erase
commands).

Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>